requirements.txt
# Core runtime dependencies
pypdf>=4.0.0
google-api-python-client>=2.130.0
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
google-generativeai>=0.7.0

# Development tooling
pytest>=8.0.0
